Implementação do CloseToGL, travado e semifuncionando

+++ src/utils/matrices.cpp
@@ -253,7 +253,7 @@ glm::mat4 Matrix_Frustum(float FOVv, float FOVh, float n, float f)
     return -M*P;
 }
 
-glm::mat4 Matrix_View_Port(int width, int height){
+glm::mat4 Matrix_View_Port(float width, float height){
     return Matrix(
         width/2,    0  , 0, width/2  , // LINHA 1
            0   , height, 0, height/2 , // LINHA 2
